# performance(stdlib): Optimise ip_cidr_contains when cidr is constant