--- a/changelog.d/1286.breaking.md
+++ b/changelog.d/1286.breaking.md
@@ -1,0 +1,38 @@
+The `ip_cidr_contains` function now validates the cidr argument during the compilation phase if it is a constant string or array. Previously, invalid constant CIDR values would only trigger an error during execution.
+
+Previous Behavior: Runtime Error
+
+Previously, if an invalid CIDR was passed as a constant, an error was thrown at runtime:
+
+```
+error[E000]: function call error for "ip_cidr_contains" at (0:45): unable to parse CIDR: couldn't parse address in network: invalid IP address syntax
+  ┌─ :1:1
+  │
+1 │ ip_cidr_contains!("INVALID", "192.168.10.32")
+  │ ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ unable to parse CIDR: couldn't parse address in network: invalid IP address syntax
+  │
+  = see language documentation at https://vrl.dev
+  = try your code in the VRL REPL, learn more at https://vrl.dev/examples
+```
+
+New Behavior: Compilation Error
+
+```
+error[E610]: function compilation error: error[E403] invalid argument
+  ┌─ :1:1
+  │
+1 │ ip_cidr_contains!("INVALID", "192.168.10.32")
+  │ ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  │ │
+  │ invalid argument "ip_cidr_contains"
+  │ error: "cidr" must be valid cidr
+  │ received: "INVALID"
+  │
+  = learn more about error code 403 at https://errors.vrl.dev/403
+  = see language documentation at https://vrl.dev
+  = try your code in the VRL REPL, learn more at https://vrl.dev/examples
+```
+
+This change improves error detection by identifying invalid CIDR values earlier, reducing unexpected failures at runtime and provides better performance.
+
+authors: JakubOnderka