--- a/src/stdlib/ip_cidr_contains.rs
+++ b/src/stdlib/ip_cidr_contains.rs
@@ -7,6 +7,20 @@ fn str_to_cidr(v: &str) -> Result<IpCidr, String> {
     IpCidr::from_str(v).map_err(|err| format!("unable to parse CIDR: {err}"))
 }
 
+fn value_to_cidr(value: &Value) -> Result<IpCidr, function::Error> {
+    let str = &value.as_str().ok_or(function::Error::InvalidArgument {
+        keyword: "ip_cidr_contains",
+        value: value.clone(),
+        error: r#""cidr" must be string"#,
+    })?;
+
+    str_to_cidr(str).map_err(|_| function::Error::InvalidArgument {
+        keyword: "ip_cidr_contains",
+        value: value.clone(),
+        error: r#""cidr" must be valid cidr"#,
+    })
+}
+
 fn ip_cidr_contains(value: &Value, cidr: &Value) -> Resolved {
     let bytes = value.try_bytes_utf8_lossy()?;
     let ip_addr =
@@ -32,6 +46,14 @@ fn ip_cidr_contains(value: &Value, cidr: &Value) -> Resolved {
         }
         .into()),
     }
+}
+
+fn ip_cidr_contains_constant(value: &Value, cidr_vec: &[IpCidr]) -> Resolved {
+    let bytes = value.try_bytes_utf8_lossy()?;
+    let ip_addr =
+        IpAddr::from_str(&bytes).map_err(|err| format!("unable to parse IP address: {err}"))?;
+
+    Ok(cidr_vec.iter().any(|cidr| cidr.contains(&ip_addr)).into())
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -70,13 +92,6 @@ impl Function for IpCidrContains {
                 result: Ok("false"),
             },
             Example {
-                title: "invalid cidr",
-                source: r#"ip_cidr_contains!("INVALID", "192.168.10.32")"#,
-                result: Err(
-                    r#"function call error for "ip_cidr_contains" at (0:45): unable to parse CIDR: couldn't parse address in network: invalid IP address syntax"#,
-                ),
-            },
-            Example {
                 title: "invalid address",
                 source: r#"ip_cidr_contains!("192.168.0.0/24", "INVALID")"#,
                 result: Err(
@@ -88,11 +103,34 @@ impl Function for IpCidrContains {
 
     fn compile(
         &self,
-        _state: &state::TypeState,
+        state: &state::TypeState,
         _ctx: &mut FunctionCompileContext,
         arguments: ArgumentList,
     ) -> Compiled {
         let cidr = arguments.required("cidr");
+
+        let cidr = match cidr.resolve_constant(state) {
+            None => IpCidrType::Expression(cidr),
+            Some(value) => IpCidrType::Constant(match value {
+                Value::Bytes(_) => vec![value_to_cidr(&value)?],
+                Value::Array(vec) => {
+                    let mut output = Vec::with_capacity(vec.len());
+                    for value in vec {
+                        output.push(value_to_cidr(&value)?);
+                    }
+                    output
+                }
+                _ => {
+                    return Err(function::Error::InvalidArgument {
+                        keyword: "ip_cidr_contains",
+                        value,
+                        error: r#""cidr" must be string or array of strings"#,
+                    }
+                    .into())
+                }
+            }),
+        };
+
         let value = arguments.required("value");
 
         Ok(IpCidrContainsFn { cidr, value }.as_expr())
@@ -100,17 +138,28 @@ impl Function for IpCidrContains {
 }
 
 #[derive(Debug, Clone)]
+enum IpCidrType {
+    Constant(Vec<IpCidr>),
+    Expression(Box<dyn Expression>),
+}
+
+#[derive(Debug, Clone)]
 struct IpCidrContainsFn {
-    cidr: Box<dyn Expression>,
+    cidr: IpCidrType,
     value: Box<dyn Expression>,
 }
 
 impl FunctionExpression for IpCidrContainsFn {
     fn resolve(&self, ctx: &mut Context) -> Resolved {
         let value = self.value.resolve(ctx)?;
-        let cidr = self.cidr.resolve(ctx)?;
 
-        ip_cidr_contains(&value, &cidr)
+        match &self.cidr {
+            IpCidrType::Constant(cidr_vec) => ip_cidr_contains_constant(&value, cidr_vec),
+            IpCidrType::Expression(exp) => {
+                let cidr = exp.resolve(ctx)?;
+                ip_cidr_contains(&value, &cidr)
+            }
+        }
     }
 
     fn type_def(&self, _: &state::TypeState) -> TypeDef {