@pront 2025-02-25
> > The change makes sense. I noticed we don't have any unit test to check errors. Also, the VRL test broke, see here https://github.com/vectordotdev/vrl/actions/runs/13487249670/job/37718273917?pr=1286
> 
> Fixed by removing failing test, because now constant invalid CIDR is validated during compilation phase. Do you consider this as breaking change?

Yes this needs a `breaking` changelog explaining what changed. I would really appreciate if we added an example (because it will help with the release changelog later).