@JakubOnderka 2025-02-27
Code style fixed